类型转换

+++ xnradmin-model/xnradmin-core/src/main/java/com/xnradmin/core/dao/mall/commodity/GoodsAllocationShowDAO.java
@@ -96,7 +96,7 @@ public class GoodsAllocationShowDAO{
         }
     }
 
-    public GoodsAllocationShow findById(Long id){
+    public GoodsAllocationShow findById(Integer id){
         log.debug("getting GoodsAllocationShow instance with id: " + id);
         try{
